tests/wpt/tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.get.order.cols.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.imageData.get.order.cols</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.imageData.get.order.cols</h1>
<p class="desc">getImageData() returns leftmost columns first</p>


<script>
var t = async_test("getImageData() returns leftmost columns first");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#fff';
  ctx.fillRect(0, 0, 100, 50);
  ctx.fillStyle = '#000';
  ctx.fillRect(0, 0, 2, 50);
  var imgdata = ctx.getImageData(0, 0, 10, 10);
  _assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
  _assertSame(imgdata.data[Math.round(imgdata.width/2*4)], 255, "imgdata.data[Math.round(imgdata.width/2*4)]", "255");
  _assertSame(imgdata.data[Math.round((imgdata.height/2)*imgdata.width*4)], 0, "imgdata.data[Math.round((imgdata.height/2)*imgdata.width*4)]", "0");
  t.done();

});
</script>
